+++ packages/graphql-codegen-factories/src/operations/__tests__/plugin.ts
@@ -1,4 +1,4 @@
-import { buildSchema, parse } from "graphql";
+import { buildSchema, parse, FragmentDefinitionNode, Kind } from "graphql";
 import { plugin } from "../plugin";
 
 describe("plugin", () => {
@@ -113,6 +113,55 @@ describe("plugin", () => {
       schema,
       [{ location: "GetMe.graphql", document: ast }],
       {}
+    );
+    expect(output).toMatchSnapshot();
+  });
+
+  it("should support external fragments", async () => {
+    const schema = buildSchema(/* GraphQL */ `
+      type User {
+        id: ID!
+        username: String!
+      }
+
+      type Query {
+        me: User!
+      }
+    `);
+
+    const fragments = parse(/* GraphQL */ `
+      fragment UserFragment on User {
+        id
+        username
+      }
+    `);
+
+    const allFragments = fragments.definitions.filter(
+      (d) => d.kind === Kind.FRAGMENT_DEFINITION
+    ) as FragmentDefinitionNode[];
+
+    const externalFragments = allFragments.map((frag) => ({
+      isExternal: true,
+      importFrom: frag.name.value,
+      name: frag.name.value,
+      onType: frag.typeCondition.name.value,
+      node: frag,
+    }));
+
+    const ast = parse(/* GraphQL */ `
+      query GetMe {
+        me {
+          ...UserFragment
+        }
+      }
+    `);
+
+    const output = await plugin(
+      schema,
+      [{ location: "GetMe.graphql", document: ast }],
+      {
+        externalFragments,
+      }
     );
     expect(output).toMatchSnapshot();
   });