+++ changelog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+## Fixed
+
+- Fix support for external fragments
+
 ## 1.0.0-beta.2 - 2022-03-27
 
 ## Fixed